Homework 7. Authorization form. Authorization with device-based token.
