+++ app/src/main/java/com/loftschool/ozaharenko/loftmoney/MainActivity.java
@@ -1,11 +1,11 @@
 package com.loftschool.ozaharenko.loftmoney;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.preference.PreferenceManager;
-import android.text.TextUtils;
 import android.view.View;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.tabs.TabLayout;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -14,21 +14,11 @@ import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
 import androidx.viewpager.widget.ViewPager;
 
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.tabs.TabLayout;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
 public class MainActivity extends AppCompatActivity {
 
     public static final String EXPENSE = "expense";
     public static final String INCOME = "income";
-    public static final String USER_ID = "ozaharenko";
     public static String TOKEN = "token";
-
-    private Api mApi;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -58,35 +48,10 @@ public class MainActivity extends AppCompatActivity {
         tabLayout.getTabAt(0).setText(R.string.expences);
         tabLayout.getTabAt(1).setText(R.string.income);
 
-        mApi = ((LoftApp)getApplication()).getApi();
-
-        final String token = PreferenceManager.getDefaultSharedPreferences(MainActivity.this).getString(TOKEN, "");
-
-        if (TextUtils.isEmpty(token)){
-
-            Call<Status> auth = mApi.auth(USER_ID);
-            auth.enqueue(new Callback<Status>() {
-                @Override
-                public void onResponse(Call<Status> call, Response<Status> response
-                ) {
-                    SharedPreferences.Editor editor = PreferenceManager.getDefaultSharedPreferences(MainActivity.this).edit();
-                    editor.putString(TOKEN, response.body().getToken());
-                    editor.apply();
-
-                    for (Fragment fragment : getSupportFragmentManager().getFragments()) {
-                        if (fragment instanceof BudgetFragment) {
-                            ((BudgetFragment)fragment).loadItems();
-                        }
-
-                    };
-                }
-
-                @Override
-                public void onFailure(Call<Status> call, Throwable t) {
-
-                }
-
-            });
+        for (Fragment fragment : getSupportFragmentManager().getFragments()) {
+            if (fragment instanceof BudgetFragment) {
+                ((BudgetFragment)fragment).loadItems();
+            }
         }
     }
