+++ app/src/main/res/values/colors.xml
@@ -5,6 +5,7 @@
     <color name="colorAccent">#ffc107</color>
     <color name="medium_grey">#666664</color>
     <color name="dark_sky_blue">#4a90e2</color>
-    <color name="white_six">#efefef</color>
     <color name="apple_green">#7ed321</color>
+    <color name="white_two">#ffffff</color>
+    <color name="white_six">#efefef</color>
 </resources>
